requirements.txt
chainlit
pyautogen
python-decouple
pandas